requirements.txt
numpy
paho-mqtt
pvlib
pandas
pyusb
requests
plotly
dash
#=======
smbus
RPi.GPIO
